ecommerce_backend/requirements.txt
Django==6.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
Pillow==11.0.0
